requirements.txt
appdirs==1.4.2
Django==1.10.6
django-cors-headers==2.0.2
djangorestframework==3.5.4
gunicorn==19.7.1
httplib2==0.10.3
magicmock==0.3
MySQL-python==1.2.5
packaging==16.8
patch==1.16
pkg-resources==0.0.0
pyparsing==2.2.0
six==1.10.0
